Added tests for higher version numbers. Interpreters should try to process the file, but if there's an interpreter error report it as something like: 'MODL Version 1 interpreter cannot process this MODL Version 100 file.'
Parser syntax errors will prevent this because the error is thrown before the interpreter has had a chance to interpret the *VERSION command. We _could_ pre-process the MODL file to check for a *VERSION command before parsing but I prefer to let the lexer and parser do their job first.